requirements.txt
open-clip-torch==2.16.0
orjson==3.8.10
packaging==23.0
pandas==2.0.0
Pillow==9.5.0
torch==2.0.0
torchvision==0.15.1
tqdm==4.65.0
gradio==3.36.1
